# Event mixing: Do not use arrow functions in groupTable