Rewrite groupTable() to use the O2 table iterators instead of manually iterating over Apache Arrow chunked columns.
Credits to @mytkom for the idea and most of the implementation.

It gives a 2x speed-up to the creation of a blocked combination policy -- see the results (mean of 5 benchmark runs) in the table below.
This does not affect the running time of creation of a combination generator and generation of combinations.

| Table size | Real time after the simplification (ns) | Real time before the simplification (ns) |
| ----------------| ----------------------------------------------------- | ---------------------------------------------------------- |
| 4 | 282.39 | 713.76
| 8 | 402.68 | 1,121.59
| 16 | 646.18 | 1,911.54
| 32 | 1,102.81 | 3,364.43
| 64 | 2,090.18 | 6,332.45
| 128 | 3,945.50 | 12,165.50
| 256 | 7,946.86 | 23,717.70
| 512 | 16,024.30 | 46,894.60
| 1024 | 32,856.40 | 95,093.30
| 2048 | 67,987.50 | 192,691.00
| 4096 | 150,429.00 | 392,947.00
| 8192 | 336,427.00 | 825,430.00
| 16384 | 716,285.00 | 1,674,110.00
| 32768 | 1,467,430.00 | 3,434,070.00
| 65536 | 3,011,900.00 | 6,886,320.00
| 131072 | 6,172,810.00 | 13,968,900.00
| 262144 | 12,517,400.00 | 28,290,700.00
| 524288 | 25,900,100.00 | 57,190,100.00
| 1048576 | 58,248,200.00 | 120,218,000.00
| 2097152 | 136,993,000.00 | 262,650,000.00

FYI @mytkom @chefxxx @MykhailoShamrai